requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
black==22.8.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1.1
django-extensions==3.2.1
django-phonenumber-field==7.0.0
django-rest==0.8.7
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.4
Faker==15.1.3
flake8==5.0.4
flower==1.2.0
humanize==4.4.0
idna==3.4
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
oauthlib==3.2.1
packaging==21.3
pathspec==0.10.1
phonenumberslite==8.12.56
Pillow==9.2.0
platformdirs==2.5.2
prometheus-client==0.15.0
prompt-toolkit==3.0.31
psycopg2-binary==2.9.5
pycodestyle==2.9.1
pycparser==2.21
pydot==1.4.2
pyflakes==2.5.0
pygraphviz-bin==0.0.1.2251946352
PyJWT==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.4
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tomli==2.0.1
tornado==6.2
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
